+++ src/path/Segment.js
@@ -3,7 +3,7 @@
  * http://paperjs.org/
  *
  * Copyright (c) 2011 - 2016, Juerg Lehni & Jonathan Puckey
- * http://scratchdisk.com/ & http://jonathanpuckey.com/
+ * http://scratchdisk.com/ & https://puckey.studio/
  *
  * Distributed under the MIT license. See LICENSE file for details.
  *
@@ -234,8 +234,8 @@ var Segment = Base.extend(/** @lends Segment# */{
      * Checks if the segment has any curve handles set.
      *
      * @return {Boolean} {@true if the segment has handles set}
-     * @see Segment#getHandleIn()
-     * @see Segment#getHandleOut()
+     * @see Segment#handleIn
+     * @see Segment#handleOut
      * @see Curve#hasHandles()
      * @see Path#hasHandles()
      */